feat: introduce initial configuration and documentation for knitwork-x
- Added `automd.config.ts` to define input files for automated documentation generation.
- Updated `README.md` to reflect the fork from knitwork, including installation instructions and usage examples for knitwork-x.
- Created new documentation structure with navigation and guides for users, including a playground for code generation examples.
- Added `.gitignore` and `package.json` for the documentation site setup.
- Included various new API documentation files for TypeScript utilities and design guidelines.
