+++ automd.config.ts
@@ -1,0 +1,3 @@
+export default {
+  input: ["README.md", "docs/1.guide/2.apis/*.md"],
+};
