+++ docs/.docs/layouts/playground.vue
@@ -1,0 +1,5 @@
+<template>
+  <UContainer>
+    <slot />
+  </UContainer>
+</template>
