+++ docs/.gitignore
@@ -1,0 +1,4 @@
+node_modules
+.nuxt
+.output
+dist
