+++ docs/1.guide/.navigation.yml
@@ -1,0 +1,1 @@
+icon: i-lucide-book-open
