+++ docs/1.guide/1.index.md
@@ -1,0 +1,37 @@
+# Getting Started
+
+Learn how to setup and use knitwork-x in your project.
+
+## Introduction
+
+knitwork-x provides utilities to generate JavaScript and TypeScript code programmatically. It is forked from [knitwork](https://github.com/unjs/knitwork) and adds comprehensive TypeScript generation helpers (classes, interfaces, functions, control flow, and more).
+
+## Installation
+
+Install knitwork-x from npm:
+
+:pm-install{name="knitwork-x"}
+
+## Usage
+
+Import the helpers you need and call them to generate code strings:
+
+```ts
+import { genImport, genClass, genConstructor } from 'knitwork-x'
+
+// Generate an import statement
+const imp = genImport('vue', ['ref', 'computed'])
+// => import { ref, computed } from "vue";
+
+// Generate a class
+const cls = genClass('Counter', [
+  genConstructor([], ['super();'])
+], { export: true })
+// => export class Counter { constructor() { super(); } }
+```
+
+## Next steps
+
+- Try the [Playground](/playground) to explore sample code and generated output.
+- See the [design guidelines](/guide/design-guidelines) for contributing to knitwork-x.
+- Full API reference is in the [README](https://github.com/hairyf/knitwork-x#readme) on GitHub.
