+++ docs/1.guide/2.design-guidelines.md
@@ -1,8 +1,9 @@
-# gen Method Design Guidelines
-
-This document summarizes design conventions for knitwork’s `gen*` APIs—parameters, return values, naming, and related norms—for reference when adding or changing code-generation helpers.
-
 ---
+title: Design Guidelines
+description: Design conventions for knitwork-x gen* APIs
+---
+
+This document summarizes design conventions for knitwork's `gen*` APIs—parameters, return values, naming, and related norms—for reference when adding or changing code-generation helpers.
 
 ## 1. Naming
 
@@ -16,18 +17,16 @@ This document summarizes design conventions for knitwork’s `gen*` APIs—param
 - **Underscore prefix**: Helpers used only within a module and not exported use the `_gen` prefix, e.g. `_genStatement`, `_genDynamicImportAttributes`.
 - **Non-gen names**: Pure utilities (e.g. escaping, wrapping) may omit `gen`, e.g. `escapeString`, `wrapInDelimiters`.
 
----
-
 ## 2. Parameters
 
 ### 2.1 Parameter order (recommended)
 
-1. **Required “subject” parameters** (what to generate): e.g. `specifier`, `name`, `object`, `members`.
-2. **Optional “subject” parameters**: e.g. `imports?`, `contents?`, `statements?`.
+1. **Required "subject" parameters** (what to generate): e.g. `specifier`, `name`, `object`, `members`.
+2. **Optional "subject" parameters**: e.g. `imports?`, `contents?`, `statements?`.
 3. **Options object**: `options: GenXxxOptions = {}`, second-to-last or last.
 4. **Indent**: If indent is supported, `indent = ""` as the **last** parameter.
 
-Examples:
+Examples
 
 ```ts
 // subject + options + indent
@@ -51,7 +50,7 @@ genParam(p: TypeField): string
 
 ### 2.3 Polymorphic input
 
-- **Single / array**: When “one” vs “many” is clear, accept `T | T[]`, e.g. `statements?: string | string[]`, `imports?: ESMImport | ESMImport[]`; normalize to an array inside the implementation.
+- **Single / array**: When "one" vs "many" is clear, accept `T | T[]`, e.g. `statements?: string | string[]`, `imports?: ESMImport | ESMImport[]`; normalize to an array inside the implementation.
 - **Object vs array**: When the same concept has two shapes (e.g. object literal vs field array), use a union, e.g. `genTypeObject(object: TypeObject | TypeObjectField[])`, `genInterface(name, contents?: TypeObject | TypeField[])`.
 
 ### 2.4 Strings and quoting
@@ -59,22 +58,18 @@ genParam(p: TypeField): string
 - **Strings that need quoting/escaping**: Always produce them via `genString(input, options)` so options like `singleQuotes` are respected.
 - **Module specifiers / paths**: Pass as `string`; the implementation uses `genString` for the quoted output.
 
----
-
 ## 3. Return value
 
 - **Type**: Always `string`—the generated code snippet (single or multi-line).
 - **Purity**: Do not mutate inputs or cause side effects; same input and options should yield stable output.
 - **Completeness**: The returned string should be a fragment that can be spliced directly into source (may include newlines and indent), e.g. a full `import` statement, a `{}` block, or a type definition.
 
----
-
 ## 4. Documentation and comments
 
 ### 4.1 JSDoc
 
 - **Short description**: One sentence stating what is generated.
-- **`@example`**: At least one example in “call + output” form, with output indicated by `// ~> `, e.g.:
+- **`@example`**: At least one example in "call + output" form, with output indicated by `// ~> `, e.g.:
   ```js
   * genString("foo");
   * // ~> `"foo"`
@@ -86,8 +81,6 @@ genParam(p: TypeField): string
 
 - Use named interfaces for structured descriptors (e.g. `TypeField`, `FunctionOpts`, `GenInterfaceOptions`) for reuse and docs.
 - Use `type` for simple unions or mappings (e.g. `ESMImport`, `LiteralField`, `EnumMemberValue`).
-
----
 
 ## 5. Implementation habits
 
@@ -108,8 +101,6 @@ genParam(p: TypeField): string
 ### 5.4 Internal reuse
 
 - Prefer reusing existing `gen*` helpers where possible (e.g. `genConstEnum` calls `genEnum`, `genInlineTypeImport` uses `genDynamicImport`) to avoid duplication and inconsistent behavior.
-
----
 
 ## 6. Quick reference
