+++ docs/2.apis/.navigation.yml
@@ -1,0 +1,2 @@
+icon: i-lucide-book-open
+name: APIs
