+++ docs/2.apis/2.variable.md
@@ -1,0 +1,97 @@
+---
+title: Variable
+description: Variable-related APIs
+---
+
+String / variable utilities（如 `genSafeVariableName`、`genString` 等）。以下内容由 automd 从源码 JSDoc 与 `@example` 自动生成。
+
+<!-- automd:jsdocs src=../../../src/string.ts group=string headingLevel=2 -->
+
+## String
+
+### `escapeString(id)`
+
+Escape a string for use in a javascript string.
+
+**Example:**
+
+```js
+escapeString("foo'bar");
+// ~> `foo\'bar`
+
+escapeString("foo\nbar");
+// ~> `foo\nbar`
+```
+
+### `genRegExp(pattern, flags?)`
+
+Generate regex literal from pattern and flags.
+
+**Example:**
+
+```js
+genRegExp("foo");
+// ~> `/foo/`
+
+genRegExp("foo", "gi");
+// ~> `/foo/gi`
+
+genRegExp("foo\\d+");
+// ~> `/foo\d+/`
+```
+
+### `genSafeVariableName(name)`
+
+Generate a safe javascript variable name.
+
+**Example:**
+
+```js
+genSafeVariableName("valid_import");
+// ~> `valid_import`
+
+genSafeVariableName("for");
+// ~> `_for`
+
+genSafeVariableName("with space");
+// ~> `with_32space`
+```
+
+### `genString(input, options)`
+
+Generate a string with double or single quotes and handle escapes.
+
+**Example:**
+
+```js
+genString("foo");
+// ~> `"foo"`
+
+genString("foo", { singleQuotes: true });
+// ~> `'foo'`
+
+genString("foo\nbar");
+// ~> `"foo\nbar"`
+```
+
+### `genTemplateLiteral(parts)`
+
+Generate runtime template literal: `` `hello ${x}` `` (value, not type).
+
+**Example:**
+
+```js
+genTemplateLiteral(["hello ", "x"]);
+// ~> `` `hello ${x}` ``
+
+genTemplateLiteral(["prefix", "expr", "suffix"]);
+// ~> `` `prefix${expr}suffix` ``
+
+genTemplateLiteral(["", "value"]);
+// ~> `` `${value}` ``
+
+genTemplateLiteral(["text"]);
+// ~> `` `text` ``
+```
+
+<!-- /automd -->
