+++ docs/2.apis/3.function.md
@@ -1,0 +1,121 @@
+---
+title: Function
+description: Function-related APIs
+---
+
+函数相关工具（如 `genParam`、`genFunction`、`genArrowFunction` 等）。以下内容由 automd 从源码 JSDoc 与 `@example` 自动生成。
+
+<!-- automd:jsdocs src=../../../src/typescript/function.ts group=Typescript headingLevel=2 -->
+
+## Typescript
+
+### `genArrowFunction(options)`
+
+Generate arrow function: `(params) => body` or `(params) => { statements }`.
+
+**Example:**
+
+```js
+genArrowFunction({ body: "x + 1" });
+// ~> `() => x + 1`
+
+genArrowFunction({ parameters: [{ name: "x", type: "number" }], body: "x * 2" });
+// ~> `(x: number) => x * 2`
+
+genArrowFunction({ parameters: [{ name: "x" }], body: ["return x + 1;"] });
+// ~> `(x) => {\n  return x + 1;\n}`
+
+genArrowFunction({ parameters: [{ name: "x", type: "string" }], body: "x.length", returnType: "number" });
+// ~> `(x: string): number => x.length`
+
+genArrowFunction({ async: true, parameters: [{ name: "url", type: "string" }], body: ["return fetch(url);"] });
+// ~> `async (url: string) => {\n  return fetch(url);\n}`
+```
+
+### `genBlock(statements?, indent)`
+
+Generate a statement block `{ statements }`.
+
+**Example:**
+
+```js
+genBlock();
+// ~> `{}`
+
+genBlock([]);
+// ~> `{}`
+
+genBlock("return x;");
+// ~> `{\n  return x;\n}`
+
+genBlock(["return x;"]);
+// ~> `{\n  return x;\n}`
+
+genBlock(["const a = 1;", "return a;"]);
+// ~> `{\n  const a = 1;\n  return a;\n}`
+
+genBlock(["return x;"], "  ");
+// ~> `{\n    return x;\n  }`
+```
+
+### `genFunction(options, indent)`
+
+Generate typescript function declaration from Function.
+
+**Example:**
+
+```js
+genFunction({ name: "foo" });
+// ~> `function foo() {}`
+
+genFunction({ name: "foo", parameters: [{ name: "x", type: "string" }, { name: "y", type: "number", optional: true }] });
+// ~> `function foo(x: string, y?: number) {}`
+
+genFunction({ name: "id", generics: [{ name: "T" }], parameters: [{ name: "x", type: "T" }], returnType: "T", body: ["return x;"] });
+// ~> `function id<T>(x: T): T { return x; }`
+
+genFunction({ name: "foo", export: true });
+// ~> `export function foo() {}`
+```
+
+### `genMethod(options, indent)`
+
+Generate shorthand object method: `name(params) { body }` (no `function` keyword).
+
+**Example:**
+
+```js
+genMethod({ name: "foo" });
+// ~> `foo() {}`
+
+genMethod({ name: "bar", parameters: [{ name: "x", type: "string" }], body: ["return x;"] });
+// ~> `bar(x: string) {\n  return x;\n}`
+
+genMethod({ name: "add", parameters: [{ name: "a", type: "number" }, { name: "b", type: "number" }], returnType: "number", body: ["return a + b;"] });
+// ~> `add(a: number, b: number): number {\n  return a + b;\n}`
+
+genMethod({ name: "fetch", async: true, parameters: [{ name: "url", type: "string" }], body: ["return await fetch(url);"] });
+// ~> `async fetch(url: string) {\n  return await fetch(url);\n}`
+```
+
+### `genParam(p)`
+
+Generate a single function parameter string from Field.
+
+**Example:**
+
+```js
+genParam({ name: "x", type: "string" });
+// ~> `x: string`
+
+genParam({ name: "y", type: "number", optional: true });
+// ~> `y?: number`
+
+genParam({ name: "z", type: "number", default: "0" });
+// ~> `z: number = 0`
+
+genParam({ name: "a" });
+// ~> `a`
+```
+
+<!-- /automd -->
