+++ docs/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "docs",
+  "private": true,
+  "scripts": {
+    "dev": "undocs dev",
+    "build": "undocs build"
+  },
+  "dependencies": {
+    "@vueuse/core": "^14.2.0",
+    "knitwork-x": "link:..",
+    "shiki": "^3.22.0"
+  },
+  "devDependencies": {
+    "nuxt": "^3.14.0",
+    "undocs": "latest",
+    "vue": "^3.5.0"
+  }
+}
